chore: remove unecessary mod
