+++ lgn-provers/src/provers/v1/preprocessing/euclid_prover.rs
@@ -3,13 +3,13 @@ use std::collections::HashMap;
 use alloy::primitives::U256;
 use anyhow::bail;
 use anyhow::Context;
-use lgn_messages::types::v1;
-use lgn_messages::types::v1::preprocessing::db_tasks::DbBlockType;
-use lgn_messages::types::v1::preprocessing::node_type;
-use lgn_messages::types::v1::preprocessing::NodeType;
-use lgn_messages::types::v1::preprocessing::WorkerTaskType;
-use lgn_messages::types::v1::ConcreteCircuitInput;
-use lgn_messages::types::v1::ConcretePublicParameters;
+use lgn_messages::v1;
+use lgn_messages::v1::preprocessing::db_tasks::DbBlockType;
+use lgn_messages::v1::preprocessing::node_type;
+use lgn_messages::v1::preprocessing::NodeType;
+use lgn_messages::v1::preprocessing::WorkerTaskType;
+use lgn_messages::v1::ConcreteCircuitInput;
+use lgn_messages::v1::ConcretePublicParameters;
 use lgn_messages::Proof;
 use mp2_common::poseidon::empty_poseidon_hash_as_vec;
 use mp2_common::types::HashOutput;
