+++ lgn-worker/src/main.rs
@@ -24,8 +24,8 @@ use lagrange::WorkerDone;
 use lagrange::WorkerToGwRequest;
 use lagrange::WorkerToGwResponse;
 use lgn_auth::jwt::JWTAuth;
-use lgn_messages::types::Message;
-use lgn_messages::types::MessageReplyEnvelope;
+use lgn_messages::Message;
+use lgn_messages::MessageReplyEnvelope;
 use lgn_worker::avs::utils::read_keystore;
 use metrics::counter;
 use mimalloc::MiMalloc;
