+++ lgn-worker/src/manager.rs
@@ -5,11 +5,11 @@ use std::panic::UnwindSafe;
 use anyhow::bail;
 use anyhow::ensure;
 use anyhow::Context;
-use lgn_messages::types::Message;
-use lgn_messages::types::MessageReplyEnvelope;
-use lgn_messages::types::ProverType;
-use lgn_messages::types::TaskDifficulty;
-use lgn_messages::types::ToProverType;
+use lgn_messages::Message;
+use lgn_messages::MessageReplyEnvelope;
+use lgn_messages::ProverType;
+use lgn_messages::TaskDifficulty;
+use lgn_messages::ToProverType;
 use lgn_provers::provers::v1::V1Prover;
 use metrics::counter;
 use metrics::histogram;
