+++ lgn-worker/src/task_difficulty.rs
@@ -1,0 +1,41 @@
+use std::fmt::Display;
+use std::fmt::Formatter;
+
+use serde::Deserialize;
+use serde::Serialize;
+
+/// The segregation of job types according to their computational complexity
+#[derive(Debug, Copy, Clone, Eq, PartialEq, PartialOrd, Ord, Serialize, Deserialize)]
+#[serde(rename_all = "lowercase")]
+pub enum TaskDifficulty {
+    // Due to the implicit ordering on which PartialOrd is built, this **MUST**
+    // remain the smaller value at the top of the enum.
+    // Hence, all workers of this class will always test .LT. *all* the tasks in
+    // queue.
+    /// Accept no tasks
+    Disabled,
+    /// Accept S tasks
+    Small,
+    /// Accept M tasks
+    Medium,
+    /// Accept L tasks
+    Large,
+}
+
+impl Display for TaskDifficulty {
+    fn fmt(
+        &self,
+        f: &mut Formatter<'_>,
+    ) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                TaskDifficulty::Small => "small",
+                TaskDifficulty::Medium => "medium",
+                TaskDifficulty::Large => "large",
+                TaskDifficulty::Disabled => "disabled",
+            }
+        )
+    }
+}
